chore: apply machine-applicable clippy fixes across the workspace (#19)
`cargo clippy --fix` over the tree (needless borrows, index loops → iterators,
redundant casts/closures, derivable/Default impls, collapsed let-chains,
excessive-precision literals, unused imports) — 70 files, behavior-preserving.
Plus the `crdt` loro deprecation: `get_or_create_container(k, LoroText::new())`
→ `ensure_mergeable_text(k)`.

Does NOT clear the advisory clippy job fully — correctness-level lints
(approx_constant, erasing_op, bad_bit_mask in DSP) and pre-existing broken
--all-targets example builds remain; tracked separately for review.

Verified: cargo check --workspace --exclude vox-discover is clean.

+++ features/fx/reverb/reverb-dsp/src/algorithms/cloud.rs
@@ -172,7 +172,7 @@ struct CloudChannel {
 impl CloudChannel {
     fn new(sample_rate: f64, is_right: bool) -> Self {
         let lines: Vec<ReverbLine> = (0..TOTAL_LINE_COUNT)
-            .map(|_| ReverbLine::new(sample_rate as f64))
+            .map(|_| ReverbLine::new(sample_rate))
             .collect();
 
         let mut diffuser = AllpassDiffuser::new_default();
